3/pyproject.toml
[project]
name = "plan-execute-agent"
version = "0.1.0"
description = "Simple Plan-Execute agent using LangGraph for learning purposes"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "pinecone>=5.0.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "requests>=2.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
